requirements.txt
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
cmake==3.18.2.post1
decorator==4.4.2
dlib==19.18.0
face-recognition==1.3.0
face-recognition-models==0.3.0
idna
imageio
imageio-ffmpeg
moviepy
numpy
opencv-python
Pillow
proglog
requests
tqdm
urllib3
wincertstore
